changed datatype to double

+++ Calculator.java
@@ -1,17 +1,17 @@
 public class Calculator {
-    public static int add(int a, int b) {
+    public static double add(double a, double b) {
         return a + b;
     }
 
-    public static int sub(int a, int b) {
+    public static double sub(double a, double b) {
         return a - b;
     }
 
-    public static int multiply(int a, int b) {
+    public static double multiply(double a, double b) {
         return a * b;
     }
 
-    public static int divide(int a, int b) {
+    public static double divide(double a, double b) {
         return a / b;
     }
 }
